Changed contentBuffer implementation
Initially, the content buffer was implemented as an array of char codes, to be re-drawn when `systemColor()` is used to reset the color of the entire window. This proved to be prohibitively slow, as everything had to be re-drawn with `tint()`, so the buffer was changed to a p5.Graphics object containing only foreground characters in white, allowing it could be re-drawn to the main graphics object in any color.

However, with this solution it was necessary to use `erase()` as each character is drawn, to revert that area of the content buffer back to transparent before drawing over it with a new character. These `erase()` calls were very slow and render performance dropped dramatically. The solution was to revert contentBuffer to a simple array of char codes, as the use of `tint()` had been eliminated since that initial implementation, allowing the entire window to be redrawn easily within one frame!