Don't allow using ESC to end `getline()`, don't move to newline
In the original projects, user input would sometimes end with the cursor on a new line, and sometimes not. Usually it doesn't matter, as the cursor is moved somewhere else right after, but for consistency, this emulator's `getline()` will not print a newline after input, leaving it to the calling code to do so if desired.