Added backup to ensure player bank balance is always valid
In C++ the bank balance was stored in an unsigned long, guaranteeing it is always a positive integer. In JS there is no such primitive type, so as a backup in case of errors in any of the games, the bank is always returned to a positive integer at the end of each game before saving.