Consolidated each project's style.css into a single file, fixed overflow
By adding `overflow: hidden;` to the `<main>` element, scrollbars are preventing from appearing and disappearing as the user zooms in or out on a page. This was a result of p5's `windowWidth/Height` rounding the root element's size instead of flooring, occasionally causing the canvas to be a fraction of a pixel larger than the window.